forth push

+++ greengoals/src/app/admincomponents/addproduct.jsx
@@ -1,10 +1,10 @@
-import { useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 
 export const ProductsPage = () => {
   const [products, setProducts] = useState([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
-  const [editingProduct, setEditingProduct] = useState(null); // For editing products
+  const [editingProductId, setEditingProductId] = useState(null);
 
   const [newProduct, setNewProduct] = useState({
     name: "",
@@ -38,35 +38,44 @@ export const ProductsPage = () => {
     e.preventDefault();
     const response = await fetch("/api/admin/product", {
       method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
       body: JSON.stringify(newProduct),
     });
     if (response.ok) {
       alert("Product added successfully!");
-      // Reload products after adding
-      const updatedProducts = await response.json();
-      setProducts((prevProducts) => [...prevProducts, updatedProducts]);
+      const updatedProduct = await response.json();
+      setProducts((prevProducts) => [...prevProducts, updatedProduct]);
+      setNewProduct({
+        name: "",
+        description: "",
+        price: "",
+        stock: "",
+        category: "",
+        images: "",
+      });
     }
   };
-  const handleEditProduct = async (e, updatedProduct) => {
-    e.preventDefault();
-    const response = await fetch(`/api/admin/product/${updatedProduct._id}`, {
+
+  const handleEditProduct = async (productId, updatedFields) => {
+    const response = await fetch(`/api/admin/product/${productId}`, {
       method: "PUT",
       headers: {
         "Content-Type": "application/json",
       },
-      body: JSON.stringify(updatedProduct),
+      body: JSON.stringify(updatedFields),
     });
 
     if (response.ok) {
       const data = await response.json();
       alert(data.message);
-      // Update the local state
       setProducts((prevProducts) =>
         prevProducts.map((product) =>
-          product._id === updatedProduct._id ? data.product : product
+          product._id === productId ? { ...product, ...updatedFields } : product
         )
       );
-      setEditingProduct(null); // Clear editing state
+      setEditingProductId(null);
     } else {
       const errorData = await response.json();
       alert(errorData.message || "Failed to update product");
@@ -74,13 +83,25 @@ export const ProductsPage = () => {
   };
 
   const handleDeleteProduct = async (productId) => {
-    const response = await fetch(`/api/admin/product/${productId}`, {
-      method: "DELETE",
-    });
-    if (response.ok) {
-      alert("Product deleted successfully!");
-      // Remove the deleted product from the state
-      setProducts(products.filter((product) => product._id !== productId));
+    if (window.confirm("Are you sure you want to delete this product?")) {
+      try {
+        const response = await fetch(`/api/admin/product/${productId}`, {
+          method: "DELETE",
+        });
+
+        if (!response.ok) {
+          throw new Error(`Error: ${response.statusText}`);
+        }
+
+        const data = await response.json();
+        alert(data.message);
+        setProducts((prevProducts) =>
+          prevProducts.filter((product) => product._id !== productId)
+        );
+      } catch (error) {
+        console.error("Failed to delete product:", error);
+        alert("Failed to delete product");
+      }
     }
   };
 
@@ -95,7 +116,6 @@ export const ProductsPage = () => {
       </div>
     );
 
-  // Group products by category
   const productsByCategory = products.reduce((acc, product) => {
     if (!acc[product.category]) {
       acc[product.category] = [];
@@ -105,167 +125,187 @@ export const ProductsPage = () => {
   }, {});
 
   return (
-    <div>
-      <h1>Products</h1>
-      <form onSubmit={handleAddProduct}>
-        <input
-          type="text"
-          name="name"
-          placeholder="Product Name"
-          onChange={handleInputChange}
-        />
-        <textarea
-          name="description"
-          placeholder="Description"
-          onChange={handleInputChange}
-        ></textarea>
-        <input
-          type="number"
-          name="price"
-          placeholder="Price"
-          onChange={handleInputChange}
-        />
-        <input
-          type="number"
-          name="stock"
-          placeholder="Stock"
-          onChange={handleInputChange}
-        />
-        <input
-          type="text"
-          name="category"
-          placeholder="Category"
-          onChange={handleInputChange}
-        />
-        <input
-          type="text"
-          name="images"
-          placeholder="Image URLs (comma-separated)"
-          onChange={handleInputChange}
-        />
-        <button type="submit">Add Product</button>
-      </form>
+    <div className="min-h-screen p-6">
+      <h1 className="text-4xl font-bold mb-8 text-center text-[#674636]">
+        Manage Products
+      </h1>
 
-      {/* Edit Form */}
-      {editingProduct && (
-        <form onSubmit={(e) => handleEditProduct(e, editingProduct)}>
+      <form
+        onSubmit={handleAddProduct}
+        className="mb-12 bg-white p-6 rounded-lg shadow-md"
+      >
+        <h2 className="text-2xl font-semibold mb-4 text-[#674636]">
+          Add New Product
+        </h2>
+        <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
           <input
             type="text"
             name="name"
-            value={editingProduct.name}
-            onChange={(e) =>
-              setEditingProduct({ ...editingProduct, name: e.target.value })
-            }
+            placeholder="Product Name"
+            value={newProduct.name}
+            onChange={handleInputChange}
+            className="w-full p-2 border border-gray-300 rounded"
           />
-          <textarea
-            name="description"
-            value={editingProduct.description}
-            onChange={(e) =>
-              setEditingProduct({
-                ...editingProduct,
-                description: e.target.value,
-              })
-            }
-          ></textarea>
           <input
             type="number"
             name="price"
-            value={editingProduct.price}
-            onChange={(e) =>
-              setEditingProduct({ ...editingProduct, price: e.target.value })
-            }
+            placeholder="Price"
+            value={newProduct.price}
+            onChange={handleInputChange}
+            className="w-full p-2 border border-gray-300 rounded"
           />
           <input
             type="number"
             name="stock"
-            value={editingProduct.stock}
-            onChange={(e) =>
-              setEditingProduct({ ...editingProduct, stock: e.target.value })
-            }
+            placeholder="Available Stock"
+            value={newProduct.stock}
+            onChange={handleInputChange}
+            className="w-full p-2 border border-gray-300 rounded"
           />
           <input
             type="text"
             name="category"
-            value={editingProduct.category}
-            onChange={(e) =>
-              setEditingProduct({ ...editingProduct, category: e.target.value })
-            }
+            placeholder="Category"
+            value={newProduct.category}
+            onChange={handleInputChange}
+            className="w-full p-2 border border-gray-300 rounded"
           />
+          <textarea
+            name="description"
+            placeholder="Product Description"
+            value={newProduct.description}
+            onChange={handleInputChange}
+            className="w-full p-2 border border-gray-300 rounded col-span-2"
+          ></textarea>
           <input
             type="text"
             name="images"
-            value={editingProduct.images}
-            onChange={(e) =>
-              setEditingProduct({ ...editingProduct, images: e.target.value })
-            }
+            placeholder="Image URLs (comma-separated)"
+            value={newProduct.images}
+            onChange={handleInputChange}
+            className="w-full p-2 border border-gray-300 rounded col-span-2"
           />
-          <button type="submit">Save Changes</button>
-        </form>
-      )}
-
-      <div className="bg-[#FFF8E8] min-h-screen">
-        <div className="container mx-auto px-4 py-8">
-          <h1 className="text-4xl font-bold mb-8 text-center text-[#674636]">
-            Product Listing
-          </h1>
-
-          {Object.entries(productsByCategory).map(
-            ([category, categoryProducts]) => (
-              <div key={category} className="mb-12">
-                <h2 className="text-3xl font-semibold mb-6 text-[#674636] border-b-2 border-[#674636] pb-2">
-                  {category}
-                </h2>
-                <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-8 justify-items-center">
-                  {categoryProducts.map((product) => (
-                    <div
-                      key={product._id}
-                      className="w-full max-w-sm bg-[#F7EED3] rounded-lg p-6 shadow-lg h-full flex flex-col justify-between"
-                    >
-                      <div>
-                        <img
-                          src={product.images[0]}
-                          alt={product.name}
-                          className="w-full h-48 object-contain rounded-md mb-4"
-                        />
-                        <h3 className="text-xl font-semibold text-[#674636] mb-2">
-                          {product.name}
-                        </h3>
-                        <p className="text-[#674636] opacity-80 mb-3 flex-grow">
-                          {product.description}
-                        </p>
-                      </div>
-                      <div className="mt-auto">
-                        <div className="flex justify-between items-center">
-                          <p className="text-lg font-bold text-[#674636]">
-                            ${product.price}
-                          </p>
-                          <p className="text-sm text-[#674636] opacity-70">
-                            Stock: {product.stock}
-                          </p>
-                        </div>
-                        <div className="flex justify-between mt-4">
-                          <button
-                            className="bg-red-500 text-white px-3 py-1 rounded"
-                            onClick={() => handleDeleteProduct(product._id)}
-                          >
-                            Delete
-                          </button>
-                          <button
-                            className="bg-blue-500 text-white px-3 py-1 rounded"
-                            onClick={() => setEditingProduct(product)}
-                          >
-                            Edit
-                          </button>
-                        </div>
-                      </div>
-                    </div>
-                  ))}
-                </div>
-              </div>
-            )
-          )}
         </div>
-      </div>
+        <button
+          type="submit"
+          className="mt-4 bg-[#674636] text-white px-4 py-2 rounded hover:bg-[#8B5E3C] transition-colors"
+        >
+          Add Product
+        </button>
+      </form>
+
+      {Object.entries(productsByCategory).map(
+        ([category, categoryProducts]) => (
+          <div key={category} className="mb-12">
+            <h2 className="text-3xl font-semibold mb-6 text-[#674636] border-b-2 border-[#674636] pb-2">
+              {category}
+            </h2>
+            <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-8">
+              {categoryProducts.map((product) => (
+                <div
+                  key={product._id}
+                  className="bg-white rounded-lg p-6 shadow-lg flex flex-col justify-between"
+                >
+                  <div>
+                    <img
+                      src={product.images[0]}
+                      alt={product.name}
+                      className="w-full h-48 object-cover rounded-md mb-4"
+                    />
+                    {editingProductId === product._id ? (
+                      <input
+                        type="text"
+                        value={product.name}
+                        onChange={(e) =>
+                          handleEditProduct(product._id, {
+                            name: e.target.value,
+                          })
+                        }
+                        className="text-xl font-semibold text-[#674636] mb-2 w-full border-b border-[#674636]"
+                      />
+                    ) : (
+                      <h3 className="text-xl font-semibold text-[#674636] mb-2">
+                        {product.name}
+                      </h3>
+                    )}
+                    {editingProductId === product._id ? (
+                      <textarea
+                        value={product.description}
+                        onChange={(e) =>
+                          handleEditProduct(product._id, {
+                            description: e.target.value,
+                          })
+                        }
+                        className="text-[#674636] opacity-80 mb-3 w-full border border-[#674636] rounded p-1"
+                      />
+                    ) : (
+                      <p className="text-[#674636] opacity-80 mb-3">
+                        {product.description}
+                      </p>
+                    )}
+                  </div>
+                  <div className="mt-auto">
+                    <div className="flex justify-between items-center mb-2">
+                      {editingProductId === product._id ? (
+                        <input
+                          type="number"
+                          value={product.price}
+                          onChange={(e) =>
+                            handleEditProduct(product._id, {
+                              price: e.target.value,
+                            })
+                          }
+                          className="text-lg font-bold text-[#674636] w-1/2 border-b border-[#674636]"
+                        />
+                      ) : (
+                        <p className="text-lg font-bold text-[#674636]">
+                          ${product.price}
+                        </p>
+                      )}
+                      {editingProductId === product._id ? (
+                        <input
+                          type="number"
+                          value={product.stock}
+                          onChange={(e) =>
+                            handleEditProduct(product._id, {
+                              stock: e.target.value,
+                            })
+                          }
+                          className="text-sm text-[#674636] opacity-70 w-1/2 border-b border-[#674636]"
+                        />
+                      ) : (
+                        <p className="text-sm text-[#674636] opacity-70">
+                          Stock: {product.stock}
+                        </p>
+                      )}
+                    </div>
+                    <div className="flex justify-between mt-4">
+                      <button
+                        className="bg-red-500 text-white px-3 py-1 rounded hover:bg-red-600 transition-colors"
+                        onClick={() => handleDeleteProduct(product._id)}
+                      >
+                        Delete
+                      </button>
+                      <button
+                        className="bg-blue-500 text-white px-3 py-1 rounded hover:bg-blue-600 transition-colors"
+                        onClick={() => {
+                          if (editingProductId === product._id) {
+                            setEditingProductId(null);
+                          } else {
+                            setEditingProductId(product._id);
+                          }
+                        }}
+                      >
+                        {editingProductId === product._id ? "Cancel" : "Edit"}
+                      </button>
+                    </div>
+                  </div>
+                </div>
+              ))}
+            </div>
+          </div>
+        )
+      )}
     </div>
   );
 };
